Add multi-line cron-style rules
There are two ways to use multi-line commands for cron-style rules.

First: If you end the cron command with a backslash, it will continue on
the next line. For example:
```
* * * * * echo \
hello \
world!
```

Second: If the cron command is just `{`, then the subsequent lines will
be executed as a script (just like with normal rules). Then a line with
just `}` will end the script. For example:
```
* * * * * {
echo hello
echo world
echo goodbye
}
```